Version 0.93-11
- Fixed GUI
- changed Memory data source

Reason:

It looks like Linux allocates memory (if it is not in use) to cache often used
files. The problem is that this memory is not realy 'used' as it will be deleted
if a process needs new memory. But sysinfo counts it as used, so we have to
parse /mem/meminfo by ourself.

Signed-off-by: Marius Messerschmidt <marius.messerschmidt@googlemail.com>